Getting rid of the dropbox code because I don't want to deal with refresh tokens and all that
