+++ chain/src/reading.py
@@ -1,8 +1,9 @@
 from langchain_core.documents import Document 
-from langchain.document_loaders import PyPDFLoader
+from langchain_community.document_loaders import PyPDFLoader
 
-from general.books import Book 
+from general.books import Book, neo_colonialism
 from general.paths import set_paths
+from general.cleaning import remove_new_line_marker
 
 
 def read_books(books: list[Book]) -> list[Document]:
@@ -11,7 +12,15 @@ def read_books(books: list[Book]) -> list[Document]:
     for book in books:
         book.download()
         loader = PyPDFLoader(file_path=str(book.file_path))
+        
+        breakpoint()
+
         loader_list.extend(loader.load())
+    
+    
+
 
     return loader_list 
 
+
+read_books([neo_colonialism])
