+++ from_scratch/pyproject.toml
@@ -19,7 +19,6 @@ sentencepiece = "^0.2.0"
 pydantic-settings = "^2.6.1"
 sentence-transformers = "^3.3.1"
 numpy = "1.26.4"
-general = {path = "../general"}
 
 [build-system]
 requires = ["poetry-core"]
