+++ from_scratch/src/reading.py
@@ -34,19 +34,6 @@ def read_pdf(book: Book) -> pymupdf.Document:
     return pymupdf.open(filename=book.file_path)
 
 
-def remove_new_line_marker(text: str) -> str:
-    """
-    Remove the commonly used new line marker "\n" from the raw text 
-
-    Args:
-        text: the raw text 
-
-    Returns:
-        str: the text with the new line marker removed
-    """
-    return text.replace("\n", " ").strip()
-
-
 def merge_books(books: list[Book], from_scratch: bool, general: bool) -> str:
     """
     Extract the raw text for each of the provided books, take the raw strings that constitute their contents,
