+++ general/general/books.py
@@ -2,9 +2,7 @@ import requests
 from pathlib import Path
 
 from loguru import logger
-from dropbox import Dropbox
 
-from general.config import general_config
 from general.paths import set_paths, make_data_directories
 
 
@@ -22,7 +20,6 @@ class Book:
 
     def download(self) -> None:
         
-        box = DropboxAPI(book=self)
         if Path(self.file_path).exists():
             logger.success(f'"{self.title}" is already saved to disk')
         else:
@@ -39,12 +36,7 @@ class Book:
                
             except Exception as error:
                 logger.error(error)
-                logger.error(f"Unable to download {self.title} from the original source.")
-                logger.warning(f'Attempting to download "{self.title}" from Dropbox')
-                box.download_from_dropbox() 
-
-        if box.book_on_dropbox():
-            box.upload_to_dropbox()
+                logger.error(f"Unable to download {self.title}.")
 
     def __find_non_core_pages__(self) -> tuple[int, int]: 
         
@@ -56,43 +48,6 @@ class Book:
         
         assert self.file_name in book_and_non_core_pages.keys()
         return book_and_non_core_pages[self.file_name]
-
-
-class DropboxAPI:
-    def __init__(self, book: Book) -> None:
-        self.book: Book = book
-        self.remote_file_path: str = f"/{book.file_name}.pdf"
-        self.api: Dropbox = Dropbox(general_config.dropbox_access_token)
-
-    def book_on_dropbox(self) -> bool:
-
-        logger.info(f'Checking whether {self.book.title} has already been uploaded')
-        metadata = self.api.files_get_metadata(path=self.remote_file_path) 
-
-        if self.remote_file_path in metadata.values():
-            logger.warning(f'"{self.book.title}" is already on Dropbox')
-            return True
-        else:
-            logger.warning(f"{self.book.title} is not on Dropbox")
-            return False 
-
-    def upload_to_dropbox(self) -> None:
-
-        logger.info(f'Attempting to upload "{self.book.title}" to Dropbox')
-
-        with open(self.book.file_path, "rb") as file:
-           self.api.files_upload(f=file.read(), path=self.remote_file_path)
-
-        logger.success(f"'{self.book.title}' successfully uploaded to Dropbox")
-
-
-    def download_from_dropbox(self) -> None:
-        metadata, response = self.api.files_download(f"/{self.book.file_name}.pdf")
-
-        with open(self.book.file_path, "wb") as file:
-            file.write(response.content)
-
-        logger.success(f"'{self.book.title}' successfully downloaded from Dropbox")
 
 
 neo_colonialism = Book(
