+++ general/pyproject.toml
@@ -10,7 +10,6 @@ python = "<3.13, >=3.9"
 requests = "^2.32.3"
 loguru = "^0.7.3"
 pydantic-settings = "^2.6.1"
-dropbox = "^12.0.2"
 python-dotenv = "^1.0.1"
 
 [build-system]
